compiler: Move IterationTree to ir/iet/utils.py
